App.c 	IRIX64 port
Base.c,Base.h,BaseI.h,BaseP.h
	Added _NhlCBobjChildChange, and changed name of _NhlCBresValueSet
	to _NhlCBobjValueSet.

Callbacks.c
	Modified Iterate function to take cbdata.

ContourPlot.c
	IRIX64 port - If line label colors seem to be messed up, then
		I made a mistake here...

Convert.c
	IRIX64

Create.c
	changes for _NhlCBobjChildChange

DataComm.c
	Corrected an error message.

Destroy.c
	_NhlCBobjChildChange

Draw.c
	Support of NhlNvpOn resource.

ERROR.f,Error.c,Format.c,GetValues.c,IrregularPlot.c,IrregularTransObj.c,
IrregularType2TransObj.c,Legend.c,LogLinPlot.c,LogLinTransObj.c,MapPlot.c,
MapTransObj.c,MultiText.c,NcgmWorkstation.c,NresDB.c,Palette.c,
StreamlinePlot.c,VectorPlot.c,Workspace.c,Workstation.c,Workstation.h,
XWorkstation.c,XyPlot.c
	IRIX64

Reparent.c
	_NhlCBobjChildChange

ResList.c
	IRIX64 - and add RLLong functions.

Resources.c,ResourcesP.h,SetValues.c,Create.c,VarArg.c,hlu.c,hluP.h
	IRIX64 (Real changes for the porting was here.)

View.c,View.h,ViewP.h
	NhlNvpOn


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@4253 10fe2890-8a40-4696-8b1d-7ccbb19c398e

+++ ni/src/lib/hlu/Convert.c
@@ -1,5 +1,5 @@
 /*
- *      $Id: Convert.c,v 1.15 1996-05-10 09:31:53 boote Exp $
+ *      $Id: Convert.c,v 1.16 1997-01-17 18:57:22 boote Exp $
  */
 /************************************************************************
 *									*
@@ -1886,7 +1886,7 @@ ComputeArgs
 		args[i].size = conv->args[i].size;
 		if(conv->args[i].addressmode == NhlLAYEROFFSET){
 			args[i].data.ptrval = (char*)ref +
-					(unsigned int)conv->args[i].data.ptrval;
+				(unsigned long)conv->args[i].data.ptrval;
 		}
 		else
 			args[i].data = conv->args[i].data;
